+++ CHANGELOG.md
@@ -7,6 +7,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org).
 
 
+[v0.3.4]
+--------
+
+### Added
+- Expire (cancel) refresh jobs that aren't executed before the data value's
+  expected expiration
+
+
 [v0.3.3]
 --------